#304 feat(service-details): disable editing if service is being updated
If the version of the service is the same or smaller after updating a service, it means the service is still being updated and editing should be disabled.
This change will disable all input fields in such a case and show an alert stating the service is currently being updated.