Skip errors while decoding config data
When docker configs are used to store binary data, `b64DecodeUnicode` will fail, the error bubbels up, and the result is that none of the configs is displayed. A single binary config will break the whole view, and no config whatsoever will be displayed. 
This commit fixes this by catching decoding errors, logging them, and displaying a note instead of the decoded data. Let me know if you require a change in the wording on line 25, or if I should use a logging facility other than `console.error`.