fix(extension): extend JWT auth token expiration for extension EE-3065
The default expiration time of 8 hours does not make sense in the
context of the docker desktop extension. This adds a new feature flag
which can be enabled with `--feat "docker-extension"` and when present
will set the expiration time to 99 years.

I've set this flag in the docker-compose.yml we use when building our
docker extension.